src/requirements.txt
flask
requests
opencv-python
imutils
Pillow
